django/profiles_project/requirements.txt
asgiref==3.2.10
astroid==2.4.2
Django==3.0.8
django-ckeditor==5.9.0
django-js-asset==1.2.2
djangorestframework==3.6.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
pylint==2.5.3
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
